+++ lib/src/public_api.ts
@@ -27,6 +27,7 @@ export * from './list/public_api';
 export * from './options/public_api';
 export * from './tabs/public_api';
 export * from './dialog/public_api';
+export * from './toast/public_api';
 export * from './angular/public_api';
 
 export * from './services/public_api';
@@ -36,3 +37,4 @@ export * from './types/public_api';
 export * from './dato-core.module';
 export * from './rx/public_api';
 export * from './errors';
+export * from './internal/deep-merge';